requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
paho-mqtt==2.1.0
PyYAML==6.0.2
